+++ pkg/datasource/getRegions.go
@@ -16,8 +16,7 @@ func (ds *Datasource) getRegions(rw http.ResponseWriter, req *http.Request) {
   }
 
   pluginConfig := httpadapter.PluginConfigFromContext(req.Context())
-  // getRegions should not require region itself
-  ec2Client, err := ds.ec2ClientFactory(&pluginConfig, "")
+  ec2Client, err := ds.ec2ClientFactory(&pluginConfig)
 
   if err != nil {
     sendError(rw, err)